requirements.txt
langchain
langchain-chroma
langchain_openai
langchain-community 
langchain-ollama
numpy
pathspec
streamlit
torch
python-dotenv
